requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
customtkinter>=5.2.0
tkinter-tooltip>=2.1.0
playsound>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
PyYAML>=6.0
tqdm>=4.65.0
psutil>=5.9.0
